task: refactor html transform to extract markdown transform
