+++ packages/cli/src/templates/app.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en" prefix="og:http://ogp.me/ns#">
+  <head>
+    <title></title>
+    <meta charset='utf-8'>
+    <meta name='viewport' content='width=device-width, initial-scale=1'/>
+    <meta-outlet></meta-outlet>
+  </head>
+  <body>
+    <section>
+      <h1>Welcome to my website!</h1>
+      <content-outlet></content-outlet>
+    </section>
+  </body>
+</html>
